Improve API by using Ok to signal success instead of a boolean
A `Result` already represents success and failure so there is no need for a
third state by including a boolean in the Result.  Additionally when an
iptables command returns a bad status it almost always includes an error
message that explains the failure.  This failure message should be
returned as an `Err` in the result.  These changes make error handling
much cleaner and failures much easier to debug.